Refactor GPS sensor to individual module.
